refactor: 모니터링 카드 컴포넌트화
